+++ Assets/Study/MessageHandler.cs
@@ -77,7 +77,7 @@ public partial class MessageHandler: UnitySingleton<MessageHandler> {
             }
         }
         //暂时不用
-        //EventCenter.Broadcast(GameEventEnum.UserEvent_NotifyHPInfo, pMsg);
+        //EventCenter.Broadcast((Int32)GameEventEnum.UserEvent_NotifyHPInfo, pMsg);
         return (Int32)EErrorCode.eNormal;  
     }
 
@@ -97,7 +97,7 @@ public partial class MessageHandler: UnitySingleton<MessageHandler> {
                 //playerXueTiao.UpdateMp();
             }
         }
-        EventCenter.Broadcast(GameEventEnum.UserEvent_NotifyMPInfo, pMsg);
+        EventCenter.Broadcast((Int32)GameEventEnum.UserEvent_NotifyMPInfo, pMsg);
         return (Int32)EErrorCode.eNormal;  
     }
 
@@ -144,7 +144,7 @@ public partial class MessageHandler: UnitySingleton<MessageHandler> {
         //普通追踪特效
         yield return 1;
         FlyEffect effect = EffectManager.Instance.CreateFlyEffect(skillPlayerID, skillTargetID, pMsg.effectid, (uint) pMsg.uniqueid, pos, dir, pMsg.ifAbsorbSkill);
-        // EventCenter.Broadcast(GameEventEnum.UserEvent_NotifySkillModelEmit, pMsg);//暂时没用上
+        // EventCenter.Broadcast((Int32)GameEventEnum.UserEvent_NotifySkillModelEmit, pMsg);//暂时没用上
     }
 
     public Int32 OnNotifySkillModelHitTarget(HitTar pMsg)
@@ -161,7 +161,7 @@ public partial class MessageHandler: UnitySingleton<MessageHandler> {
         UInt64 targetID;
         targetID = pMsg.targuid;
 
-        EventCenter.Broadcast<UInt64, uint, UInt64>(GameEventEnum.GameEvent_BroadcastBeAtk, ownerID, pMsg.effectid, targetID);//添加警告  光圈
+        EventCenter.Broadcast<UInt64, uint, UInt64>((Int32)GameEventEnum.GameEvent_BroadcastBeAtk, ownerID, pMsg.effectid, targetID);//添加警告  光圈
         yield return 1;
         HolyTech.Effect.EffectManager.Instance.CreateBeAttackEffect(ownerID, targetID, pMsg.effectid);//创建受击特效
     }
@@ -246,18 +246,18 @@ public partial class MessageHandler: UnitySingleton<MessageHandler> {
             //释放技能 
              entity.OnEntityReleaseSkill();  
         }   
-        EventCenter.Broadcast(GameEventEnum.UserEvent_NotifyGameObjectReleaseSkillState, pMsg);
+        EventCenter.Broadcast((Int32)GameEventEnum.UserEvent_NotifyGameObjectReleaseSkillState, pMsg);
         return (int)EErrorCode.eNormal;
     }
 
     public int  OnNotifySkillInfo(NotifySkillInfo pMsg){
-       EventCenter.Broadcast(GameEventEnum.UserEvent_NotifySkillInfo, pMsg);
+       EventCenter.Broadcast((Int32)GameEventEnum.UserEvent_NotifySkillInfo, pMsg);
        return (int)EErrorCode.eNormal;
    }
 
     public int OnNotifySGameObjectFreeState(FreeState pMsg  )
     {
-        EventCenter.Broadcast(GameEventEnum.UserEvent_NotifyGameObjectFreeState, pMsg);
+        EventCenter.Broadcast((Int32)GameEventEnum.UserEvent_NotifyGameObjectFreeState, pMsg);
         return (int)EErrorCode.eNormal;
     }
 
@@ -270,13 +270,13 @@ public partial class MessageHandler: UnitySingleton<MessageHandler> {
         entity.GOSSI.fBeginTime = Time.realtimeSinceStartup;
         entity.GOSSI.fLastSyncSecond = Time.realtimeSinceStartup;
         
-        EventCenter.Broadcast(GameEventEnum.UserEvent_NotifyGameObjectRunState,pMsg);
+        EventCenter.Broadcast((Int32)GameEventEnum.UserEvent_NotifyGameObjectRunState,pMsg);
         return (int)EErrorCode.eNormal;
     }
 
     public int OnNotifySkillModelStartForceMoveTeleport(NotifySkillModelStartForceMoveTeleport pMsg)
     {
-        EventCenter.Broadcast(GameEventEnum.UserEvent_NotifySkillModelStartForceMoveTeleport, pMsg);
+        EventCenter.Broadcast((Int32)GameEventEnum.UserEvent_NotifySkillModelStartForceMoveTeleport, pMsg);
         return (int)EErrorCode.eNormal; 
     }
 
@@ -284,7 +284,7 @@ public partial class MessageHandler: UnitySingleton<MessageHandler> {
     {
         //玩家确定英雄 显示加载界面 
         //HeroCtrl.Instance.AddRealSelectHero((uint)pMsg.heroposinfo.pos, pMsg.heroposinfo.heroid);
-        EventCenter.Broadcast(GameEventEnum.UserEvent_NotifyEnsureHero, pMsg);
+        EventCenter.Broadcast((Int32)GameEventEnum.UserEvent_NotifyEnsureHero, pMsg);
         return (int)EErrorCode.eNormal;
     }
 
@@ -296,20 +296,20 @@ public partial class MessageHandler: UnitySingleton<MessageHandler> {
 
     public int OnNotifyGameObjectAppear(GSToGC.GOAppear pMsg)
     {    
-        EventCenter.Broadcast(GameEventEnum.UserEvent_NotifyGameObjectAppear, pMsg);  
+        EventCenter.Broadcast((Int32)GameEventEnum.UserEvent_NotifyGameObjectAppear, pMsg);  
         return (int)EErrorCode.eNormal;
     }
 
     public int OnBroadcastBattleHeroInfo(GSToGC.BroadcastBattleHeroInfo pMsg)
     {
         //英雄Id
-        EventCenter.Broadcast(GameEventEnum.UserEvent_NotifyBattleHeroInfo, pMsg);
+        EventCenter.Broadcast((Int32)GameEventEnum.UserEvent_NotifyBattleHeroInfo, pMsg);
         return (int)EErrorCode.eNormal;
     }
 
     public int OnNotifyToChooseHero(GSToGC.TryToChooseHero pMsg)
     {
-        EventCenter.Broadcast(GameEventEnum.UserEvent_NotifyTryChooseHero, pMsg);
+        EventCenter.Broadcast((Int32)GameEventEnum.UserEvent_NotifyTryChooseHero, pMsg);
         return (int)EErrorCode.eNormal;
     }
 
@@ -323,7 +323,7 @@ public partial class MessageHandler: UnitySingleton<MessageHandler> {
         {
             GameUserModel.Instance.GameBattleID = pMsg.battleid;
             GameUserModel.Instance.GameMapID = (uint)pMsg.mapid;
-            EventCenter.Broadcast(GameEventEnum.GameEvent_ReconnectToBatttle);
+            EventCenter.Broadcast((Int32)GameEventEnum.GameEvent_ReconnectToBatttle);
         }
         else
         {
@@ -336,13 +336,13 @@ public partial class MessageHandler: UnitySingleton<MessageHandler> {
     {
 
 
-        EventCenter.Broadcast(GameEventEnum.UserEvent_NotifyBattleSeatPosInfo, pMsg);
+        EventCenter.Broadcast((Int32)GameEventEnum.UserEvent_NotifyBattleSeatPosInfo, pMsg);
         return (int)EErrorCode.eNormal;
     }
 
     public int OnNotifyBattleMatherCount(GSToGC.BattleMatcherCount pMsg)
     {
-        EventCenter.Broadcast(GameEventEnum.UserEvent_NotifyBattleMatherCount, pMsg);
+        EventCenter.Broadcast((Int32)GameEventEnum.UserEvent_NotifyBattleMatherCount, pMsg);
         return (int)EErrorCode.eNormal;
     }
 
@@ -367,7 +367,7 @@ public partial class MessageHandler: UnitySingleton<MessageHandler> {
 
     public int OnNotifyMatchTeamSwitch(GSToGC.NotifyMatchTeamSwitch pMsg)
     {
-        EventCenter.Broadcast(GameEventEnum.UserEvent_NotifyMatchTeamSwitch, pMsg.startflag);
+        EventCenter.Broadcast((Int32)GameEventEnum.UserEvent_NotifyMatchTeamSwitch, pMsg.startflag);
         return (int)EErrorCode.eNormal;
     }
 
@@ -375,7 +375,7 @@ public partial class MessageHandler: UnitySingleton<MessageHandler> {
     {
         //初始化组队伍基本信息  设置队伍地图id 及匹配模式
         //TeamMatchCtrl.Instance.InitTeamBaseInfo(pMsg.mapid, pMsg.matchtype);
-        EventCenter.Broadcast<bool>(GameEventEnum.UserEvent_NotifyMatchTeamBaseInfo, pMsg.teamid!=0);
+        EventCenter.Broadcast<bool>((Int32)GameEventEnum.UserEvent_NotifyMatchTeamBaseInfo, pMsg.teamid!=0);
         return (int)EErrorCode.eNormal;
     }
 
@@ -411,7 +411,7 @@ public partial class MessageHandler: UnitySingleton<MessageHandler> {
             SelectServerData.Instance.SetServerList(i, name, (SelectServerData.ServerState)state, addr, port, area);
         }
 
-        EventCenter.Broadcast(GameEventEnum.UserEvent_NotifyServerAddr);
+        EventCenter.Broadcast((Int32)GameEventEnum.UserEvent_NotifyServerAddr);
         return (int)EErrorCode.eNormal;
     }
 
@@ -422,7 +422,7 @@ public partial class MessageHandler: UnitySingleton<MessageHandler> {
         SelectServerData.Instance.GateServerPort = pMsg.port;
         SelectServerData.Instance.GateServerToken = pMsg.token;
         SelectServerData.Instance.SetGateServerUin(pMsg.user_name);
-        EventCenter.Broadcast(GameEventEnum.UserEvent_NotifyGateServerInfo, pMsg);
+        EventCenter.Broadcast((Int32)GameEventEnum.UserEvent_NotifyGateServerInfo, pMsg);
         return (int)EErrorCode.eNormal;
     }
 
@@ -440,7 +440,7 @@ public partial class MessageHandler: UnitySingleton<MessageHandler> {
     public int OnNotifyBattleStateChange(Stream stream)
     {
         //这个消息用于显示选择选择英雄界面 所以在显示窗口的位置注册
-      //  EventCenter.Broadcast(GameEventEnum.GameEvent_IntoHero);
+      //  EventCenter.Broadcast((Int32)GameEventEnum.GameEvent_IntoHero);
         return (int)EErrorCode.eNormal;
     }
 
@@ -452,13 +452,13 @@ public partial class MessageHandler: UnitySingleton<MessageHandler> {
             GameUserModel.Instance.CanChooseHeroList.Add(heroId);
         }
         GameUserModel.Instance.STCTimeDiff = pMsg.timeDiff;
-        EventCenter.Broadcast(GameEventEnum.UserEvent_NotifyHeroList, pMsg);
+        EventCenter.Broadcast((Int32)GameEventEnum.UserEvent_NotifyHeroList, pMsg);
         return (int)EErrorCode.eNormal;
     }
 
     public int OnNotifyBattleStateChange(BattleStateChange pMsg)
     {
-        EventCenter.Broadcast(GameEventEnum.UserEvent_NotifyBattleStateChange, pMsg);
+        EventCenter.Broadcast((Int32)GameEventEnum.UserEvent_NotifyBattleStateChange, pMsg);
         return (int)EErrorCode.eNormal;
     }
